--- functions.js
@@ -1,1 +1,0 @@
-let getPokemonById = () => {}
