+++ public/editpokeform.handlebars
@@ -3,13 +3,14 @@
 
 <head>
     <meta charset="UTF-8">
-    <title>Edit Pokemon Time</title>
+    <title>Welcome To Pokedex!</title>
 </head>
 
 <body>
-    <form method="POST" action="/pokemon/edit/">
+    <form method="POST" action="/pokemon/edit">
+        <h1>Which pokemon ID do you want to edit?</h1>
         <div>
-            <label>Which pokemon ID do you want to edit?</label>
+            <label>ID</label>
             <input type="text" name="id">
         </div>
         <div>
