+++ public/home.handlebars
@@ -9,9 +9,10 @@
 <body>
     <form method="GET" action="/">
         <input type="hidden" name="sortby" value="name">
+        <h1>Welcome To Pokedex!</h1>
         <div>
-            <label>Sort by name</label>
-            <input type="submit" value="Submit">
+            <label>Sort the pokemons by name</label>
+            <input type="submit" value="Sort">
         </div>
     </form>
     <ul>
