+++ public/newpokeform.handlebars
@@ -3,11 +3,12 @@
 
 <head>
     <meta charset="UTF-8">
-    <title>Add New Pokemon</title>
+    <title>Welcome To Pokedex!</title>
 </head>
 
 <body>
-    <form method="POST" action="/pokemon/new/">
+    <form method="POST" action="/pokemon">
+        <h1>What pokemon do you want to create?</h1>
         <div>
             <label>Name</label>
             <input type="text" name="name">
